requirements.txt
bcrypt             4.0.1
capstone           5.0.1
certifi            2023.7.22
cffi               1.15.1
charset-normalizer 3.2.0
colored-traceback  0.3.0
cryptography       41.0.4
gmpy2              2.1.5
idna               3.4
intervaltree       3.1.0
Mako               1.2.4
MarkupSafe         2.1.3
packaging          23.1
paramiko           3.3.1
pip                23.2.1
plumbum            1.8.2
psutil             5.9.5
pwntools           4.11.0
pycparser          2.21
pycryptodome       3.19.0
pyelftools         0.30
Pygments           2.16.1
PySocks            1.7.1
python-dateutil    2.8.2
pywin32            306
requests           2.31.0
ROPGadget          7.4
rpyc               5.3.1
setuptools         68.2.0
six                1.16.0
sortedcontainers   2.4.0
unicorn            2.0.1.post1
urllib3            2.0.5
wheel              0.41.2